requirements.txt
opencv-python==4.8.0.74
tensorflow==2.12.0
numpy==1.23.5
pillow==9.5.0
matplotlib==3.7.1
scikit-learn==1.2.2 